+++ changes.md
@@ -2,6 +2,6 @@
 
 * small error in layerswitch
 * fixed gestures (sgpx did not work because of this)
-* cluster now works with leaflet-marker, leaflet-gpx, leaflet-geojson
+* cluster now works with leaflet-marker, leaflet-gpx, leaflet-geojson, leaflet-kml
 * multielavation: it works with multiple maps on one page now.
 * WordPress 5.9